% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emparam.R
\name{emparam}
\alias{emparam}
\title{EM Algorithm Estimator for Parametric Model}
\usage{
emparam(d_vec, w_vec, cutoff, x_dist = c("gauss"), u_dist = c("gauss",
  "lap"), reltol = 1e-06, maxit = 1000L, integ_method = c("romberg",
  "simpson", "trapezoid", "simpson2"), integ_reltol = 1e-06,
  integ_depth = 100L, verbose = FALSE, ...)
}
\arguments{
\item{d_vec}{binary integer vector of assignment}

\item{w_vec}{numeric vector of observed running variable}

\item{cutoff}{threshold value for assignment}

\item{reltol}{relative tolerance requied}

\item{maxit}{maximum number of iteration}

\item{integ_method, integ_reltol, integ_depth}{control parameters for numerical integration}

\item{verbose}{if true, progress is reported}

\item{...}{currently not used}
}
\description{
EM Algorithm Estimator for Parametric Model
}
\examples{
\dontrun{
dat <- gen_data(500, 0.2, 0)
emparam(dat$d, dat$w, 0, x_dist = "gauss", u_dist = "lap", verbose = TRUE)
}
}

