% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_data}
\alias{gen_data}
\title{Generate Simulation Data for a Sharp RDD with Mismeasured Running Variable}
\usage{
gen_data(n, sigma, cutoff, u_dist = c("gauss", "lap"), x_dist = c("gauss",
  "exp", "unif", "gaussmix"), mu_x = 0, sd_x = 1, rate_x = 1,
  min_x = -sqrt(3), max_x = sqrt(3), nmix = 5, mu_min = -1.5,
  mu_max = 1.5, sds = 0.25, ...)
}
\arguments{
\item{n}{sample size}

\item{sigma}{standard deviation of the measurement error term}

\item{cutoff}{threshold value for the assignment}

\item{u_dist}{distribution of the measurement error.
Only "gauss" or "lap" is supported.}

\item{x_dist}{distribution of the true running variable.
Either "gauss", "gaussmix" or any function that a generate random sample.}

\item{mu_x, sd_x}{mean and standard deviation of gaussian distribution.
Used if \code{x_dist} is 'gauss'}

\item{rate_x}{rate parameter used when \code{x_dist} is 'exp'}

\item{min_x, max_x}{min and max parameter used when \code{x_dist} is 'unif'}

\item{nmix, mu_min, mu_max, sds}{parameters used
when \code{x_dist} is 'gaussmix'.
\code{nmix} is the number of mixture. \code{mu_min} and \code{mu_max} is
the range of means. \code{sds} is the standard deviations.}

\item{...}{additional argument to be passed to the generator function for
running variable}
}
\value{
list of 4 numeric vectors of length n.
\itemize{
\item{d}: assignment variable
\item{w}: observed running variable
\item{x}: true running variable
\item{u}: measurement error
}
}
\description{
Generate Simulation Data for a Sharp RDD with Mismeasured Running Variable
}
\examples{
gen_data(100, 0.2, 0)
gen_data(100, 0.2, 1, u_dist = "lap", x_dist = "exp")
}

