% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsgauss.R
\name{tsgauss}
\alias{tsgauss}
\title{Two Step Gaussian Model for Measurement Error Variance for a Sharp RDD}
\usage{
tsgauss(d_vec, w_vec, ...)
}
\arguments{
\item{d_vec}{binary integer vector of assignment}

\item{w_vec}{numeric vector of observed running variable}
}
\value{
estimate for the standard deviation
}
\description{
Two Step Gaussian Model for Measurement Error Variance for a Sharp RDD
}

