% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsgauss.R
\name{tsgauss}
\alias{tsgauss}
\title{Two Step Gaussian Estimator}
\usage{
tsgauss(d_vec, w_vec, cutoff, init_ratio = 0.25, ...)
}
\arguments{
\item{d_vec}{binary integer vector of assignment}

\item{w_vec}{numeric vector of observed running variable}

\item{cutoff}{threshold value for assignment}

\item{init_ratio}{ratio of the intial sigma^2
to the sample variance of \code{w}. the numeric between 0 and 1}

\item{...}{additional controls for \code{optim}}
}
\value{
object of \code{rddsigma} class
}
\description{
Estimate the standard deviation of measurement error in the
running variable of sharp RDD.  This estimator is constructed under
the assumption that true running variable and measurement error
follow the Gaussian distribution.
}
\examples{
dat <- gen_data(500, 0.2, 0)
tsgauss(dat$d, dat$w, 0)
}
\references{
Kevin M. Murphy and Robert H. Topel (1985), Estimation and Inference in Two-Step Econometric Models. Journal of Business & Economic Statistics, 3(4), pp.370-379
}

